Fixed some typo Fixxed residual underscore names in some file (delitem.php, delete.php, edit.php etc.) Changed example logic: before, example files used the classes in 'lib' folder but this made me implement wrong logic for config file. Now in example folder you find a FormProcessor folder which contains classes used by the example files (like normally we'd have in real world.

+++ example/delitem.php
@@ -2,10 +2,10 @@
 
 error_reporting(E_ALL);
 ini_set('display_errors', 'on');
-require_once '../lib/FormModel.class.php';
-$config = require '../lib/config.php';
-$root = getenv('DOCUMENT_ROOT').DIRECTORY_SEPARATOR;
-$root_target_dir = $config['general']['rootTargetDir'];
+require_once 'FormProcessor/FormModel.class.php';
+$config = require 'FormProcessor/config.php';
+$root = getenv('DOCUMENT_ROOT') . DIRECTORY_SEPARATOR;
+$rootTargetDir = $config['general']['rootTargetDir'];
 $itemId = filter_input(INPUT_POST, 'itemId', FILTER_SANITIZE_NUMBER_INT);
 $confirm = filter_input(INPUT_POST, 'confirm');
 if (isset($confirm) && isset($itemId)) {
@@ -16,17 +16,21 @@ if (isset($confirm) && isset($itemId)) {
     $pics = explode(';', $pictures);
     if (count($docs) > 0) {
         foreach ($docs as $d) {
-            if (file_exists($root.$rootTargetDir.'documents/'.$d)) {
-                echo $root.$rootTargetDir.'documents/'.$d.' EXISTS<br>';
-                unlink($root.$rootTargetDir.'documents/'.$d);
+            if (!empty($d)) {
+                if (file_exists($root . $rootTargetDir . 'documents/' . $d)) {
+                    echo $root . $rootTargetDir . 'documents/' . $d . ' EXISTS<br>';
+                    unlink($root . $rootTargetDir . 'documents/' . $d);
+                }
             }
         }
     }
     if (count($pics) > 0) {
         foreach ($pics as $p) {
-            if (file_exists($root.$rootTargetDir.'pictures/'.$p)) {
-                echo $root.$rootTargetDir.'pictures/'.$p.' EXISTS<br>';
-                unlink($root.$rootTargetDir.'pictures/'.$p);
+            if (!empty($p)) {
+                if (file_exists($root . $rootTargetDir . 'pictures/' . $p)) {
+                    echo $root . $rootTargetDir . 'pictures/' . $p . ' EXISTS<br>';
+                    unlink($root . $rootTargetDir . 'pictures/' . $p);
+                }
             }
         }
     }
